requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.2
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==3.0.3
django-filter==2.3.0
django-import-export==2.4.0
django-rest-swagger==2.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
gunicorn==19.8.1
idna==2.10
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rest-framework-generic-relations==2.0.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
